requirements.txt
aiogram==2.23.1
aiohttp==3.8.6
aioredis==2.0.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
exceptiongroup==1.2.0
frozenlist==1.4.1
idna==3.6
iniconfig==2.0.0
lxml==4.9.4
magic-filter==1.0.12
multidict==6.0.4
packaging==23.2
pluggy==1.3.0
pyTelegramBotAPI==4.7.1
pytest==7.4.3
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
soupsieve==2.5
telegram==0.0.1
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.1.0
yarl==1.9.4
